+++ src/haven/Party.java
@@ -30,7 +30,7 @@ import java.util.*;
 import java.awt.Color;
 
 public class Party {
-    Map<Integer, Member> memb = new TreeMap<Integer, Member>();
+    Map<Long, Member> memb = new TreeMap<Long, Member>();
     Member leader = null;
     public static final int PD_LIST = 0;
     public static final int PD_LEADER = 1;
@@ -42,7 +42,7 @@ public class Party {
     }
 	
     public class Member {
-	int gobid;
+	long gobid;
 	private Coord c = null;
 	Color col = Color.BLACK;
 	
@@ -62,15 +62,15 @@ public class Party {
 	while(!msg.eom()) {
 	    int type = msg.uint8();
 	    if(type == PD_LIST) {
-		ArrayList<Integer> ids = new ArrayList<Integer>();
+		ArrayList<Long> ids = new ArrayList<Long>();
 		while(true) {
-		    int id = msg.int32();
+		    long id = msg.int32();
 		    if(id < 0)
 			break;
 		    ids.add(id);
 		}
-		Map<Integer, Member> nmemb = new TreeMap<Integer, Member>();
-		for(int id : ids) {
+		Map<Long, Member> nmemb = new TreeMap<Long, Member>();
+		for(long id : ids) {
 		    Member m = memb.get(id);
 		    if(m == null) {
 			m = new Member();
@@ -78,15 +78,15 @@ public class Party {
 		    }
 		    nmemb.put(id, m);
 		}
-		int lid = (leader == null)?-1:leader.gobid;
+		long lid = (leader == null)?-1:leader.gobid;
 		memb = nmemb;
 		leader = memb.get(lid);
 	    } else if(type == PD_LEADER) {
-		Member m = memb.get(msg.int32());
+		Member m = memb.get((long)msg.int32());
 		if(m != null)
 		    leader = m;
 	    } else if(type == PD_MEMBER) {
-		Member m = memb.get(msg.int32());
+		Member m = memb.get((long)msg.int32());
 		Coord c = null;
 		boolean vis = msg.uint8() == 1;
 		if(vis)
